+++ packages/headless/src/ssr-commerce-next.index.ts
@@ -97,6 +97,7 @@ export type {
 } from './controllers/controller/headless-controller.js';
 export type {
   BreadcrumbManager,
+  BreadcrumbManagerDefinition,
   BreadcrumbManagerState,
 } from './ssr-next/commerce/controllers/breadcrumb-manager/headless-core-breadcrumb-manager.ssr.js';
 export {defineBreadcrumbManager} from './ssr-next/commerce/controllers/breadcrumb-manager/headless-core-breadcrumb-manager.ssr.js';
@@ -122,6 +123,7 @@ export type {
 export {defineContext} from './ssr-next/commerce/controllers/context/headless-context.ssr.js';
 export type {
   DidYouMean,
+  DidYouMeanDefinition,
   DidYouMeanState,
 } from './ssr-next/commerce/controllers/did-you-mean/headless-did-you-mean.ssr.js';
 export {defineDidYouMean} from './ssr-next/commerce/controllers/did-you-mean/headless-did-you-mean.ssr.js';
@@ -147,6 +149,7 @@ export type {
   DateFacetState,
   DateFacetValue,
   FacetGenerator,
+  FacetGeneratorDefinition,
   FacetGeneratorState,
   FacetType,
   LocationFacet,
@@ -174,6 +177,7 @@ export {defineInstantProducts} from './ssr-next/commerce/controllers/instant-pro
 export type {
   CorePaginationOptions,
   Pagination,
+  PaginationDefinition,
   PaginationOptions,
   PaginationProps,
   PaginationState,
@@ -192,12 +196,16 @@ export type {
 export {defineParameterManager} from './ssr-next/commerce/controllers/parameter-manager/headless-core-parameter-manager.ssr.js';
 export type {
   ProductList,
+  ProductListDefinition,
   ProductListing,
   ProductListState,
   Search,
 } from './ssr-next/commerce/controllers/product-list/headless-product-list.ssr.js';
 export {defineProductList} from './ssr-next/commerce/controllers/product-list/headless-product-list.ssr.js';
-export type {ProductView} from './ssr-next/commerce/controllers/product-view/headless-product-view.ssr.js';
+export type {
+  ProductView,
+  ProductViewDefinition,
+} from './ssr-next/commerce/controllers/product-view/headless-product-view.ssr.js';
 export {defineProductView} from './ssr-next/commerce/controllers/product-view/headless-product-view.ssr.js';
 export type {
   RecentQueriesList,
@@ -226,6 +234,7 @@ export type {
 export {defineSearchBox} from './ssr-next/commerce/controllers/search-box/headless-search-box.ssr.js';
 export type {
   Sort,
+  SortDefinition,
   SortProps,
   SortState,
 } from './ssr-next/commerce/controllers/sort/headless-core-commerce-sort.ssr.js';
@@ -242,6 +251,7 @@ export type {
   RecommendationsSummaryState,
   SearchSummaryState,
   Summary,
+  SummaryDefinition,
   SummaryState,
 } from './ssr-next/commerce/controllers/summary/headless-core-summary.ssr.js';
 export {defineSummary} from './ssr-next/commerce/controllers/summary/headless-core-summary.ssr.js';
@@ -264,10 +274,7 @@ export type {
 } from './ssr-next/commerce/controllers/triggers/headless-redirection-trigger.ssr.js';
 export {defineRedirectionTrigger} from './ssr-next/commerce/controllers/triggers/headless-redirection-trigger.ssr.js';
 export {defineCommerceEngine} from './ssr-next/commerce/engine/commerce-engine.ssr.js';
-export type {
-  CommerceEngineDefinitionOptions,
-  SSRCommerceEngine as CommerceEngine,
-} from './ssr-next/commerce/factories/build-factory.js';
+export type {SSRCommerceEngine as CommerceEngine} from './ssr-next/commerce/factories/build-factory.js';
 // Main App
 export type {
   BuildConfig,
@@ -275,10 +282,9 @@ export type {
 } from './ssr-next/commerce/types/build.js';
 export {SolutionType} from './ssr-next/commerce/types/controller-constants.js';
 export type {
+  CommerceEngineDefinitionControllersPropsOption as EngineDefinitionControllersPropsOption,
   ControllerDefinition,
   ControllerDefinitionsMap,
-  ControllerWithKind,
-  EngineDefinitionControllersPropsOption,
   HydratedState,
   InferControllerStaticStateFromController,
   InferControllerStaticStateMapFromControllers,
@@ -295,12 +301,12 @@ export type {
 export type {
   CommerceEngineDefinition,
   CommerceEngineDefinitionBuildResult,
+  CommerceEngineDefinitionOptions,
   EngineStaticState,
   FetchStaticState,
   HydrateStaticState,
   HydrateStaticStateOptions,
 } from './ssr-next/commerce/types/engine.js';
-export {Kind} from './ssr-next/commerce/types/kind.js';
 // State
 export type {
   CommerceAppState,