+++ packages/headless/src/ssr-next/commerce/controllers/context/headless-context.ssr.ts
@@ -9,7 +9,6 @@ import {
 } from '../../../../controllers/commerce/context/headless-context.js';
 import {MissingControllerProps} from '../../../common/errors.js';
 import type {UniversalControllerDefinitionWithProps} from '../../types/controller-definitions.js';
-import {createControllerWithKind, Kind} from '../../types/kind.js';
 
 export type {
   Context,
@@ -20,11 +19,10 @@ export type {
   ContextOptions,
 };
 
-export interface ContextDefinition
-  extends UniversalControllerDefinitionWithProps<
-    Context,
-    {initialState: ContextOptions}
-  > {}
+export type ContextDefinition = UniversalControllerDefinitionWithProps<
+  Context,
+  {initialState: ContextOptions}
+>;
 
 /**
  * Defines a `Context` controller instance.
@@ -40,10 +38,9 @@ export function defineContext(): ContextDefinition {
     recommendation: true,
     buildWithProps: (engine, props: {initialState: ContextOptions}) => {
       if (props === undefined) {
-        throw new MissingControllerProps(Kind.Context);
+        throw new MissingControllerProps('Context');
       }
-      const controller = buildContext(engine, {options: props.initialState});
-      return createControllerWithKind(controller, Kind.Context);
+      return buildContext(engine, {options: props.initialState});
     },
   };
 }