+++ packages/headless/src/ssr-next/commerce/types/controller-scopes.ts
@@ -41,13 +41,6 @@ export interface SearchOnlyController {
   [SolutionType.search]: true;
 }
 
-export interface ListingOnlyController {
-  /**
-   * @internal
-   */
-  [SolutionType.listing]: true;
-}
-
 export interface RecommendationOnlyController {
   /**
    * @internal