+++ packages/headless/src/ssr-next/search/controller-utils.ts
@@ -2,15 +2,15 @@ import type {UnknownAction} from '@reduxjs/toolkit';
 import type {CoreEngine, CoreEngineNext} from '../../app/engine.js';
 import type {Controller} from '../../controllers/controller/headless-controller.js';
 import {clone, mapObject} from '../../utils/utils.js';
-import {ControllerBuilder} from './builders/controller-builder.js';
-import {createStaticControllerBuilder} from './builders/static-controller-builder.js';
-import type {ControllerDefinitionsMap} from './types/controllers.js';
-import type {EngineStaticState} from './types/engine.js';
+import {ControllerBuilder} from '../common/builders/controller-builder.js';
+import {createStaticControllerBuilder} from '../common/builders/static-controller-builder.js';
+import type {EngineStaticState} from '../common/types/engine.js';
+import type {ControllerDefinitionsMap} from './types/controller-definition.js';
 import type {
   InferControllerPropsMapFromDefinitions,
   InferControllerStaticStateMapFromDefinitions,
   InferControllersMapFromDefinition,
-} from './types/inference.js';
+} from './types/controller-inference.js';
 
 export function buildControllerDefinitions<
   TControllerDefinitionsMap extends ControllerDefinitionsMap<