+++ packages/headless/src/ssr-next/search/controllers/did-you-mean/headless-did-you-mean.ssr.ts
@@ -3,7 +3,7 @@ import {
   buildDidYouMean,
   type DidYouMean,
 } from '../../../../controllers/did-you-mean/headless-did-you-mean.js';
-import type {ControllerDefinitionWithoutProps} from '../../../common/types/controllers.js';
+import type {ControllerDefinitionWithoutProps} from '../../types/controller-definition.js';
 
 export * from '../../../../controllers/did-you-mean/headless-did-you-mean.js';