+++ packages/headless/src/ssr-next/search/controllers/tab/headless-tab.ssr.ts
@@ -4,7 +4,7 @@ import {
   type Tab,
   type TabProps,
 } from '../../../../controllers/tab/headless-tab.js';
-import type {ControllerDefinitionWithoutProps} from '../../../common/types/controllers.js';
+import type {ControllerDefinitionWithoutProps} from '../../types/controller-definition.js';
 
 export * from '../../../../controllers/tab/headless-tab.js';