+++ packages/headless/src/ssr-next/search/controllers/triggers/headless-query-trigger.ssr.ts
@@ -1,7 +1,7 @@
 import type {SearchEngine} from '../../../../app/search-engine/search-engine.js';
 import type {QueryTrigger} from '../../../../controllers/core/triggers/headless-core-query-trigger.js';
 import {buildQueryTrigger} from '../../../../controllers/triggers/headless-query-trigger.js';
-import type {ControllerDefinitionWithoutProps} from '../../../common/types/controllers.js';
+import type {ControllerDefinitionWithoutProps} from '../../types/controller-definition.js';
 
 export type {
   QueryTrigger,