+++ packages/headless/src/ssr-next/search/types/build.ts
@@ -1,0 +1,23 @@
+import type {CoreEngine, CoreEngineNext} from '../../../app/engine.js';
+import type {
+  ControllersMap,
+  ControllersPropsMap,
+} from '../../common/types/controllers.js';
+import type {OptionsTuple} from '../../common/types/utilities.js';
+import type {
+  SearchEngineDefinitionBuildResult,
+  SearchEngineDefinitionControllersPropsOption,
+} from './engine.js';
+
+/**
+ * @internal
+ */
+export type Build<
+  TEngine extends CoreEngine | CoreEngineNext,
+  TControllersMap extends ControllersMap,
+  TControllersProps extends ControllersPropsMap,
+> = (
+  ...params: OptionsTuple<
+    SearchEngineDefinitionControllersPropsOption<TControllersProps>
+  >
+) => Promise<SearchEngineDefinitionBuildResult<TEngine, TControllersMap>>;